// admin/src/types/Product.ts
export interface Product {
    productId: number;
    name: string;
    description: string;
    price: number;
    stock: number;
    category: string; 
  }


  export interface ProductDetail {
    productId: number;
    name: string;
    description: string;
    price: number;
    stock: number;
    category: number; 
  }